+++ src/Exception/InvalidConfigException.php
@@ -1,12 +1,14 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-authorization for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-authorization/blob/master/LICENSE.md New BSD License
  */
 
+declare(strict_types=1);
+
 namespace Zend\Expressive\Authorization\Exception;
 
-class InvalidConfigException extends RuntimeException implements ExceptionInterface
+class InvalidConfigException extends RuntimeException
 {
 }