+++ test/AuthorizationMiddlewareFactoryTest.php
@@ -1,15 +1,20 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-authorization for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2017-2018 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-authorization/blob/master/LICENSE.md New BSD License
  */
 
+declare(strict_types=1);
+
 namespace ZendTest\Expressive\Authorization;
 
+use PHPUnit\Framework\Assert;
 use PHPUnit\Framework\TestCase;
+use Prophecy\Prophecy\ObjectProphecy;
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ResponseInterface;
+use ReflectionProperty;
 use Zend\Expressive\Authorization\AuthorizationInterface;
 use Zend\Expressive\Authorization\AuthorizationMiddleware;
 use Zend\Expressive\Authorization\AuthorizationMiddlewareFactory;
@@ -17,19 +22,37 @@ use Zend\Expressive\Authorization\Exception;
 
 class AuthorizationMiddlewareFactoryTest extends TestCase
 {
+    /** @var ContainerInterface|ObjectProphecy */
+    private $container;
+
+    /** @var AuthorizationMiddlewareFactory */
+    private $factory;
+
+    /** @var AuthorizationInterface|ObjectProphecy */
+    private $authorization;
+
+    /** @var ResponseInterface|ObjectProphecy */
+    private $responsePrototype;
+
+    /** @var callable */
+    private $responseFactory;
+
     protected function setUp()
     {
         $this->container = $this->prophesize(ContainerInterface::class);
         $this->factory = new AuthorizationMiddlewareFactory();
         $this->authorization = $this->prophesize(AuthorizationInterface::class);
-        $this->response = $this->prophesize(ResponseInterface::class);
+        $this->responsePrototype = $this->prophesize(ResponseInterface::class);
+        $this->responseFactory = function () {
+            return $this->responsePrototype->reveal();
+        };
 
         $this->container
             ->get(AuthorizationInterface::class)
             ->will([$this->authorization, 'reveal']);
         $this->container
             ->get(ResponseInterface::class)
-            ->will([$this->response, 'reveal']);
+            ->willReturn($this->responseFactory);
     }
 
     public function testFactoryWithoutAuthorization()
@@ -37,16 +60,7 @@ class AuthorizationMiddlewareFactoryTest extends TestCase
         $this->container->has(AuthorizationInterface::class)->willReturn(false);
 
         $this->expectException(Exception\InvalidConfigException::class);
-        $middleware = ($this->factory)($this->container->reveal());
-    }
-
-    public function testFactoryWithoutResponsePrototype()
-    {
-        $this->container->has(AuthorizationInterface::class)->willReturn(true);
-        $this->container->has(ResponseInterface::class)->willReturn(false);
-
-        $this->expectException(Exception\InvalidConfigException::class);
-        $middleware = ($this->factory)($this->container->reveal());
+        ($this->factory)($this->container->reveal());
     }
 
     public function testFactory()
@@ -56,5 +70,16 @@ class AuthorizationMiddlewareFactoryTest extends TestCase
 
         $middleware = ($this->factory)($this->container->reveal());
         $this->assertInstanceOf(AuthorizationMiddleware::class, $middleware);
+        $this->assertResponseFactoryReturns($this->responsePrototype->reveal(), $middleware);
+    }
+
+    public static function assertResponseFactoryReturns(
+        ResponseInterface $expected,
+        AuthorizationMiddleware $middleware
+    ) : void {
+        $r = new ReflectionProperty($middleware, 'responseFactory');
+        $r->setAccessible(true);
+        $responseFactory = $r->getValue($middleware);
+        Assert::assertSame($expected, $responseFactory());
     }
 }