+++ test/AuthorizationMiddlewareTest.php
@@ -1,59 +1,78 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-authorization for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2017-2018 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-authorization/blob/master/LICENSE.md New BSD License
  */
+
+declare(strict_types=1);
 
 namespace ZendTest\Expressive\Authorization;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
+use Prophecy\Prophecy\ObjectProphecy;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use Webimpress\HttpMiddlewareCompatibility\HandlerInterface;
+use Psr\Http\Server\RequestHandlerInterface;
 use Zend\Expressive\Authentication\UserInterface;
 use Zend\Expressive\Authentication\UserRepository\UserTrait;
 use Zend\Expressive\Authorization\AuthorizationInterface;
 use Zend\Expressive\Authorization\AuthorizationMiddleware;
 
-use const Webimpress\HttpMiddlewareCompatibility\HANDLER_METHOD;
-
 class AuthorizationMiddlewareTest extends TestCase
 {
     use UserTrait;
+
+    /** @var AuthorizationInterface|ObjectProphecy */
+    private $authorization;
+
+    /** @var ServerRequestInterface|ObjectProphecy */
+    private $request;
+
+    /** @var RequestHandlerInterface|ObjectProphecy */
+    private $handler;
+
+    /** @var ResponseInterface|ObjectProphecy */
+    private $responsePrototype;
+
+    /** @var callable */
+    private $responseFactory;
 
     protected function setUp()
     {
         $this->authorization = $this->prophesize(AuthorizationInterface::class);
         $this->request = $this->prophesize(ServerRequestInterface::class);
-        $this->delegate = $this->prophesize(HandlerInterface::class);
-        $this->response = $this->prophesize(ResponseInterface::class);
+        $this->handler = $this->prophesize(RequestHandlerInterface::class);
+        $this->responsePrototype = $this->prophesize(ResponseInterface::class);
+        $this->responseFactory = function () {
+            return $this->responsePrototype->reveal();
+        };
     }
 
     public function testConstructor()
     {
-        $middleware = new AuthorizationMiddleware($this->authorization->reveal(), $this->response->reveal());
+        $middleware = new AuthorizationMiddleware($this->authorization->reveal(), $this->responseFactory);
         $this->assertInstanceOf(AuthorizationMiddleware::class, $middleware);
     }
 
     public function testProcessWithoutUserAttribute()
     {
         $this->request->getAttribute(UserInterface::class, false)->willReturn(false);
-        $this->response->withStatus(401)->will([$this->response, 'reveal']);
+        $this->responsePrototype->withStatus(401)->will([$this->responsePrototype, 'reveal']);
 
-        $this->delegate
-            ->{HANDLER_METHOD}(Argument::any())
+        $this->handler
+            ->handle(Argument::any())
             ->shouldNotBeCalled();
 
-        $middleware = new AuthorizationMiddleware($this->authorization->reveal(), $this->response->reveal());
+        $middleware = new AuthorizationMiddleware($this->authorization->reveal(), $this->responseFactory);
 
         $response = $middleware->process(
             $this->request->reveal(),
-            $this->delegate->reveal()
+            $this->handler->reveal()
         );
 
-        $this->assertSame($this->response->reveal(), $response);
+        $this->assertSame($this->responsePrototype->reveal(), $response);
     }
 
     public function testProcessRoleNotGranted()
@@ -61,25 +80,25 @@ class AuthorizationMiddlewareTest extends TestCase
         $this->request
             ->getAttribute(UserInterface::class, false)
             ->willReturn($this->generateUser('foo', ['bar']));
-        $this->response
+        $this->responsePrototype
             ->withStatus(403)
-            ->will([$this->response, 'reveal']);
+            ->will([$this->responsePrototype, 'reveal']);
         $this->authorization
             ->isGranted('bar', Argument::that([$this->request, 'reveal']))
             ->willReturn(false);
 
-        $this->delegate
-            ->{HANDLER_METHOD}(Argument::any())
+        $this->handler
+            ->handle(Argument::any())
             ->shouldNotBeCalled();
 
-        $middleware = new AuthorizationMiddleware($this->authorization->reveal(), $this->response->reveal());
+        $middleware = new AuthorizationMiddleware($this->authorization->reveal(), $this->responseFactory);
 
         $response = $middleware->process(
             $this->request->reveal(),
-            $this->delegate->reveal()
+            $this->handler->reveal()
         );
 
-        $this->assertSame($this->response->reveal(), $response);
+        $this->assertSame($this->responsePrototype->reveal(), $response);
     }
 
     public function testProcessRoleGranted()
@@ -91,17 +110,17 @@ class AuthorizationMiddlewareTest extends TestCase
             ->isGranted('bar', Argument::that([$this->request, 'reveal']))
             ->willReturn(true);
 
-        $this->delegate
-            ->{HANDLER_METHOD}(Argument::any())
-            ->will([$this->response, 'reveal']);
+        $this->handler
+            ->handle(Argument::any())
+            ->will([$this->responsePrototype, 'reveal']);
 
-        $middleware = new AuthorizationMiddleware($this->authorization->reveal(), $this->response->reveal());
+        $middleware = new AuthorizationMiddleware($this->authorization->reveal(), $this->responseFactory);
 
         $response = $middleware->process(
             $this->request->reveal(),
-            $this->delegate->reveal()
+            $this->handler->reveal()
         );
 
-        $this->assertSame($this->response->reveal(), $response);
+        $this->assertSame($this->responsePrototype->reveal(), $response);
     }
 }